UI re-integration
New library integration finished. Project works to the standard it did last year, however, with updated code.
Voltage sensor working (although maybe 0.1v short of the true value) - turns out all I had to do was re-wire it once I'd finished the software side of things. Turns out I had connected it to the 3.3v pin on the mega instead of the 5v pin, since they're right next to each other - and I watch netflix while doing this stuff (which I really shouldn't).

Need to re-code UI and implement joystick interface so I can add the features to get this project to the final level of its development.

For now, the next item on the list is to reorganise and tidy up the code for the UI.